+++ src/com/blade2d/drawelements/QuadElem.java
@@ -7,7 +7,7 @@ public class QuadElem extends AbstractDrawElem {
 	float x1, x2, x3, x4, y1, y2, y3, y4;
 	float r, g, b, a;
 	
-	public QuadElem(int x1, int y1, int x2, int y2, int x3, int y3, int x4, int y4, float r, float g, float b, float a,
+	public QuadElem(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4, float r, float g, float b, float a,
 			int l) {
 		super(l);
 		this.x1 = x1;
